warning C4018: '>=' : signed/unsigned mismatch fix
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@4985 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ src/splittransactionsdialog.cpp
@@ -217,7 +217,7 @@ void SplitTransactionDialog::UpdateSplitTotal()
 
 void SplitTransactionDialog::EditEntry(int index)
 {
-    if (index < 0 || index >= this->m_local_splits.size()) return;
+    if (index < 0 || index >= (int)this->m_local_splits.size()) return;
     Model_Splittransaction::Data& split = this->m_local_splits[index];
     SplitDetailDialog sdd(this, &split, transType_);
     if (sdd.ShowModal() == wxID_OK)
